Make CI tests faster on commit

+++ scripts/common/__scripts-tests.sh
@@ -11,8 +11,13 @@
 #
 #      "exit" with non-zero.
 #
-WORKDIRS="$HOME/suibase/workdirs"
-OUT="$HOME/suibase/scripts/tests/result.txt"
+export WORKDIRS="$HOME/suibase/workdirs"
+export OUT="$HOME/suibase/scripts/tests/result.txt"
+
+# Initialized by test_setup()
+export FAST_OPTION=false
+export MAIN_BRANCH_OPTION=false
+export TEST_INIT_CALLED=false
 
 # List of workdir routinely tested.
 export WORKDIRS_NAME=(
@@ -79,40 +84,58 @@ assert_build_ok() {
   [ "$_FIRST_WORD" = "sui" ] || fail "sui --version did not work [$_VERSION]"
 }
 
-test_setup() {
-  # Parse command-line
-  #GITHUB_OPTION=false
-  #while [[ "$#" -gt 0 ]]; do
-  #  case $1 in
-  # -t|--target) target="$2"; shift ;; That's an example with a parameter
-  # -f|--flag) flag=1 ;; That's an example flag
-  #      --github) GITHUB_OPTION=true ;;
-  #      *)
-  #      fail "Unknown parameter passed: $1";
-  #  esac
-  #  shift
-  #done
-
-  # Clean-up from potential previous execution.
+test_init() {
+  if [ "$TEST_INIT_CALLED" = "true" ]; then
+    return
+  fi
+  TEST_INIT_CALLED=true
   rm -rf "$OUT"
-
   # This script should not be called from under workdirs since it will get deleted.
   local _USER_CWD
   _USER_CWD=$(pwd -P)
-  if [[ "$_USER_CWD" = *"suibase/workdirs"* ]]; then
+  if [[ "$_USER_CWD" == *"suibase/workdirs"* ]]; then
     fail "Should not call this test from a directory that will get deleted [suibase/workdirs]"
   fi
 
-  # Add here tests done on github.
-
-  #if [ "$GITHUB_OPTION" = true ]; then
-  # Success on github if reaching here.
-  #  echo "Test Completed (github early exit)"
-  #  exit 0
-  #fi
-
-  # Add here tests not run on github.
-  #echo "Test Completed"
 }
+export -f test_init
 
-test_setup
+test_setup() {
+
+  # Parse command-line
+  FAST_OPTION=false
+  MAIN_BRANCH_OPTION=false
+  local _SKIP_INIT=false
+  while [[ "$#" -gt 0 ]]; do
+    case $1 in
+    --fast) FAST_OPTION=true ;;
+    --main_branch) MAIN_BRANCH_OPTION=true ;;
+    --skip_init) _SKIP_INIT=true ;;
+    *)
+      fail "Unknown parameter passed: $1"
+      ;;
+    esac
+    shift
+  done
+
+  # run-all.sh calls test_init() once
+  # before calling the test sub-scripts.
+  #
+  # If caller is not run-all.sh (the subscript
+  # is called directly on the command line), then
+  # the test_init() is called from here instead.
+  #
+  if [ "$_SKIP_INIT" = "false" ]; then
+    test_init
+  else
+    # If caller says to skip init, then
+    # jsut assume it was already done.
+    TEST_INIT_CALLED=true
+  fi
+
+  # Sanity check.
+  if [ "$TEST_INIT_CALLED" = "false" ]; then
+    fail "test_init() not called"
+  fi
+}
+export -f test_setup
